chore: move pc_ic map to a new module
