+++ crates/evm/core/src/ic.rs
@@ -1,0 +1,57 @@
+use revm::{
+    interpreter::{opcode, opcode::spec_opcode_gas},
+    primitives::SpecId,
+};
+use std::collections::BTreeMap;
+
+/// A map of program counters to instruction counters.
+pub type PCICMap = BTreeMap<usize, usize>;
+
+/// Builds a mapping from program counters to instruction counters.
+pub fn build_pc_ic_map(spec: SpecId, code: &[u8]) -> PCICMap {
+    let opcode_infos = spec_opcode_gas(spec);
+    let mut pc_ic_map: PCICMap = BTreeMap::new();
+
+    let mut i = 0;
+    let mut cumulative_push_size = 0;
+    while i < code.len() {
+        let op = code[i];
+        pc_ic_map.insert(i, i - cumulative_push_size);
+        if opcode_infos[op as usize].is_push() {
+            // Skip the push bytes.
+            //
+            // For more context on the math, see: https://github.com/bluealloy/revm/blob/007b8807b5ad7705d3cacce4d92b89d880a83301/crates/revm/src/interpreter/contract.rs#L114-L115
+            i += (op - opcode::PUSH1 + 1) as usize;
+            cumulative_push_size += (op - opcode::PUSH1 + 1) as usize;
+        }
+        i += 1;
+    }
+
+    pc_ic_map
+}
+
+/// A map of instruction counters to program counters.
+pub type ICPCMap = BTreeMap<usize, usize>;
+
+/// Builds a mapping from instruction counters to program counters.
+pub fn build_ic_pc_map(spec: SpecId, code: &[u8]) -> ICPCMap {
+    let opcode_infos = spec_opcode_gas(spec);
+    let mut ic_pc_map: ICPCMap = ICPCMap::new();
+
+    let mut i = 0;
+    let mut cumulative_push_size = 0;
+    while i < code.len() {
+        let op = code[i];
+        ic_pc_map.insert(i - cumulative_push_size, i);
+        if opcode_infos[op as usize].is_push() {
+            // Skip the push bytes.
+            //
+            // For more context on the math, see: https://github.com/bluealloy/revm/blob/007b8807b5ad7705d3cacce4d92b89d880a83301/crates/revm/src/interpreter/contract.rs#L114-L115
+            i += (op - opcode::PUSH1 + 1) as usize;
+            cumulative_push_size += (op - opcode::PUSH1 + 1) as usize;
+        }
+        i += 1;
+    }
+
+    ic_pc_map
+}
