+++ crates/evm/core/src/utils.rs
@@ -1,18 +1,19 @@
-use alloy_json_abi::{Function, JsonAbi as Abi};
-use alloy_primitives::{Address, FixedBytes, B256};
+use alloy_json_abi::{Function, JsonAbi};
+use alloy_primitives::FixedBytes;
 use alloy_rpc_types::{Block, Transaction};
 use ethers::types::{ActionType, CallType, Chain, H256, U256};
 use eyre::ContextCompat;
 use foundry_common::types::ToAlloy;
-pub use foundry_compilers::utils::RuntimeOrHandle;
-pub use revm::primitives::State as StateChangeset;
-
 use revm::{
-    interpreter::{opcode, opcode::spec_opcode_gas, CallScheme, CreateInputs, InstructionResult},
+    interpreter::{CallScheme, InstructionResult},
     primitives::{CreateScheme, Eval, Halt, SpecId, TransactTo},
 };
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
+
+pub use foundry_compilers::utils::RuntimeOrHandle;
+pub use revm::primitives::State as StateChangeset;
+
+pub use crate::ic::*;
 
 // TODO(onbjerg): Remove this and use `CallKind` from the tracer.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
@@ -177,63 +178,11 @@ pub fn apply_chain_and_block_specific_env_changes(env: &mut revm::primitives::En
     }
 }
 
-/// A map of program counters to instruction counters.
-pub type PCICMap = BTreeMap<usize, usize>;
-
-/// Builds a mapping from program counters to instruction counters.
-pub fn build_pc_ic_map(spec: SpecId, code: &[u8]) -> PCICMap {
-    let opcode_infos = spec_opcode_gas(spec);
-    let mut pc_ic_map: PCICMap = BTreeMap::new();
-
-    let mut i = 0;
-    let mut cumulative_push_size = 0;
-    while i < code.len() {
-        let op = code[i];
-        pc_ic_map.insert(i, i - cumulative_push_size);
-        if opcode_infos[op as usize].is_push() {
-            // Skip the push bytes.
-            //
-            // For more context on the math, see: https://github.com/bluealloy/revm/blob/007b8807b5ad7705d3cacce4d92b89d880a83301/crates/revm/src/interpreter/contract.rs#L114-L115
-            i += (op - opcode::PUSH1 + 1) as usize;
-            cumulative_push_size += (op - opcode::PUSH1 + 1) as usize;
-        }
-        i += 1;
-    }
-
-    pc_ic_map
-}
-
-/// A map of instruction counters to program counters.
-pub type ICPCMap = BTreeMap<usize, usize>;
-
-/// Builds a mapping from instruction counters to program counters.
-pub fn build_ic_pc_map(spec: SpecId, code: &[u8]) -> ICPCMap {
-    let opcode_infos = spec_opcode_gas(spec);
-    let mut ic_pc_map: ICPCMap = ICPCMap::new();
-
-    let mut i = 0;
-    let mut cumulative_push_size = 0;
-    while i < code.len() {
-        let op = code[i];
-        ic_pc_map.insert(i - cumulative_push_size, i);
-        if opcode_infos[op as usize].is_push() {
-            // Skip the push bytes.
-            //
-            // For more context on the math, see: https://github.com/bluealloy/revm/blob/007b8807b5ad7705d3cacce4d92b89d880a83301/crates/revm/src/interpreter/contract.rs#L114-L115
-            i += (op - opcode::PUSH1 + 1) as usize;
-            cumulative_push_size += (op - opcode::PUSH1 + 1) as usize;
-        }
-        i += 1;
-    }
-
-    ic_pc_map
-}
-
 /// Given an ABI and selector, it tries to find the respective function.
 pub fn get_function(
     contract_name: &str,
     selector: &FixedBytes<4>,
-    abi: &Abi,
+    abi: &JsonAbi,
 ) -> eyre::Result<Function> {
     abi.functions()
         .find(|func| func.selector().as_slice() == selector.as_slice())
